third_party/WebKit/LayoutTests/html/dialog/inert-label-focus-expected.txt

Tests focusing of an inert label for a non-inert target. label.focus() should send focus to the target, but clicking the label should be the same as clicking on the document body.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing that calling focus() on label sends focus to its target.
PASS document.activeElement is document.querySelector('#submit')
Testing that clicking on the label sends focus to document.body.
PASS document.activeElement is document.body
PASS successfullyParsed is true

TEST COMPLETE
Label for Submit   
